Update rubocop deps (#2254)
* Update rubocop dependencies

* Update rubocop channel on .codeclimate.yml

* Run rubocop autocorrections

+++ app/workers/delete_stale_data_worker.rb
@@ -9,9 +9,7 @@ class DeleteStaleDataWorker
       .where('updated_at < ?', (Time.current - ACTIVE_DURATION))
       .where(status_id: Users::DataRequest::STATUS[:success])
       .each do |dr|
-      if dr.file_path.present? && File.exist?(dr.file_path)
-        File.delete(dr.file_path)
-      end
+      File.delete(dr.file_path) if dr.file_path.present? && File.exist?(dr.file_path)
       dr.update(status_id: Users::DataRequest::STATUS[:deleted])
     end
   end
